requirements.txt
pyserial==3.4
thingspeak==0.4.1
